# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Seat {
  id: number;
  seatNumber: number;
  /** AC or NON_AC */
  section: string;
  isAC: boolean;
  isUnderMaintenance: boolean;
  price: number;
  /**
     * True if this seat is booked for the queried month
     * @nullable
     */
  bookedForMonth?: boolean | null;
  /**
     * Booking ID if booked for the queried month
     * @nullable
     */
  bookingId?: number | null;
  /**
     * Name of the person who booked (admin view)
     * @nullable
     */
  bookedByName?: string | null;
}

export interface SeatUpdate {
  isUnderMaintenance?: boolean;
}

export interface Booking {
  id: number;
  seatId: number;
  seatNumber: number;
  section: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** YYYY-MM format */
  month: string;
  amount: number;
  /** pending | confirmed | cancelled */
  status: string;
  /** @nullable */
  paymentSessionId?: string | null;
  createdAt: string;
}

export interface BookingInput {
  seatId: number;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  customerPhone: string;
  customerEmail?: string;
  /** YYYY-MM format */
  month: string;
}

export interface BookingUpdate {
  /** pending | confirmed | cancelled */
  status?: string;
}

export interface BookingSummary {
  month: string;
  totalSeats: number;
  bookedSeats: number;
  availableSeats: number;
  acBooked: number;
  nonAcBooked: number;
  revenue: number;
}

export interface PaymentInitiate {
  bookingId: number;
}

export interface PaymentSession {
  sessionId: string;
  bookingId: number;
  amount: number;
  expiresAt: string;
}

export interface PaymentConfirm {
  sessionId: string;
  bookingId: number;
  cardNumber?: string;
  cardName?: string;
  expiry?: string;
  cvv?: string;
}

export interface Pricing {
  id: number;
  acPrice: number;
  nonAcPrice: number;
  updatedAt: string;
}

export interface PricingUpdate {
  acPrice?: number;
  nonAcPrice?: number;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  token: string;
}

export type ListSeatsParams = {
/**
 * Month in YYYY-MM format (e.g. 2026-07)
 */
month?: string;
};

export type ListBookingsParams = {
/**
 * Month in YYYY-MM format
 */
month?: string;
};

export type GetBookingSummaryParams = {
/**
 * Month in YYYY-MM format
 */
month: string;
};