# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  seatId: number;
  seatNumber: number;
  section: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** YYYY-MM format */
  month: string;
  amount: number;
  /** pending | confirmed | cancelled */
  status: string;
  /** @nullable */
  paymentSessionId?: string | null;
  createdAt: string;
}