# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  seatId: number;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  customerPhone: string;
  customerEmail?: string;
  /** YYYY-MM format */
  month: string;
}