# lib/api-zod/src/generated/types/bookingSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingSummary {
  month: string;
  totalSeats: number;
  bookedSeats: number;
  availableSeats: number;
  acBooked: number;
  nonAcBooked: number;
  revenue: number;
}