# lib/api-zod/src/generated/types/paymentInitiate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInitiate {
  bookingId: number;
}