# lib/api-zod/src/generated/types/paymentSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSession {
  sessionId: string;
  bookingId: number;
  amount: number;
  expiresAt: string;
}