# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lucky Reading Room - Library Seat Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Seat {
  id: number;
  seatNumber: number;
  /** AC or NON_AC */
  section: string;
  isAC: boolean;
  isUnderMaintenance: boolean;
  price: number;
  /**
     * True if this seat is booked for the queried month
     * @nullable
     */
  bookedForMonth?: boolean | null;
  /**
     * Booking ID if booked for the queried month
     * @nullable
     */
  bookingId?: number | null;
  /**
     * Name of the person who booked (admin view)
     * @nullable
     */
  bookedByName?: string | null;
}